arambhgpt-backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-generativeai
asyncpg
pydantic[email]
email-validator
schedule
apscheduler